+++ lib/src/visual/views/camera/camera_preview_components/camera_preview_controller_view.dart
@@ -63,23 +63,25 @@ class CameraPreviewControllerView extends StatefulWidget {
   final bool hideControllers;
 
   @override
-  State<CameraPreviewControllerView> createState() => _CameraPreviewControllerViewState();
+  State<CameraPreviewControllerView> createState() =>
+      _CameraPreviewControllerViewState();
 }
 
-class _CameraPreviewControllerViewState extends State<CameraPreviewControllerView> {
+class _CameraPreviewControllerViewState
+    extends State<CameraPreviewControllerView> {
   Future<bool>? _permissionsFuture;
 
   @override
   void initState() {
     super.initState();
-    if (!AppState.hasCameraPermissions) {
-      _permissionsFuture = checkPermissions().then((hasPermission) {
-        if (hasPermission) {
+    _permissionsFuture = checkPermissions().then((hasPermission) {
+      if (hasPermission && mounted) {
+        setState(() {
           AppState.hasCameraPermissions = true;
-        }
-        return hasPermission;
-      });
-    }
+        });
+      }
+      return hasPermission;
+    });
   }
 
   @override
@@ -107,6 +109,10 @@ class _CameraPreviewControllerViewState extends State<CameraPreviewControllerVie
           } else {
             return PermissionHandlerView(
               onSuccess: () {
+                setState(() {
+                  AppState.hasCameraPermissions = true;
+                  _permissionsFuture = Future.value(true);
+                });
                 widget.mainController.selectCamera(0, true);
               },
             );
@@ -241,7 +247,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
   Future<void> initAsync() async {
     _hasAudioPermission = await Permission.microphone.isGranted;
 
-    if (!_hasAudioPermission && !userService.currentUser.requestedAudioPermission) {
+    if (!_hasAudioPermission &&
+        !userService.currentUser.requestedAudioPermission) {
       await UserService.update((u) => u.requestedAudioPermission = true);
       await requestMicrophonePermission();
     }
@@ -262,7 +269,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
   }
 
   Future<void> updateScaleFactor(double newScale) async {
-    if (mc.selectedCameraDetails.scaleFactor == newScale || mc.cameraController == null) {
+    if (mc.selectedCameraDetails.scaleFactor == newScale ||
+        mc.cameraController == null) {
       return;
     }
     await mc.cameraController?.setZoomLevel(
@@ -345,7 +353,9 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
     bool sharedFromGallery = false,
     MediaType? mediaType,
   }) async {
-    final type = mediaType ?? ((videoFilePath != null) ? MediaType.video : MediaType.image);
+    final type =
+        mediaType ??
+        ((videoFilePath != null) ? MediaType.video : MediaType.image);
     final mediaFileService = await initializeMediaUpload(
       type,
       userService.currentUser.defaultShowTime,
@@ -386,9 +396,10 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
                   mainCameraController: mc,
                   previewLink: mc.sharedLinkForPreview,
                 ),
-                transitionsBuilder: (context, animation, secondaryAnimation, child) {
-                  return child;
-                },
+                transitionsBuilder:
+                    (context, animation, secondaryAnimation, child) {
+                      return child;
+                    },
                 transitionDuration: Duration.zero,
                 reverseTransitionDuration: Duration.zero,
               ),
@@ -418,13 +429,16 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
     return false;
   }
 
-  bool get isFront => mc.cameraController?.description.lensDirection == CameraLensDirection.front;
+  bool get isFront =>
+      mc.cameraController?.description.lensDirection ==
+      CameraLensDirection.front;
 
   Future<void> onPanUpdate(dynamic details) async {
     if (details == null) {
       return;
     }
-    if (mc.cameraController == null || !mc.cameraController!.value.isInitialized) {
+    if (mc.cameraController == null ||
+        !mc.cameraController!.value.isInitialized) {
       return;
     }
 
@@ -553,7 +567,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
   }
 
   Future<void> startVideoRecording() async {
-    if (mc.cameraController != null && mc.cameraController!.value.isRecordingVideo) {
+    if (mc.cameraController != null &&
+        mc.cameraController!.value.isRecordingVideo) {
       return;
     }
     setState(() {
@@ -573,7 +588,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
           _currentTime = clock.now();
         });
         if (_videoRecordingStarted != null &&
-            _currentTime.difference(_videoRecordingStarted!).inSeconds >= maxVideoRecordingTime) {
+            _currentTime.difference(_videoRecordingStarted!).inSeconds >=
+                maxVideoRecordingTime) {
           timer.cancel();
           _videoRecordingTimer = null;
           stopVideoRecording();
@@ -610,7 +626,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
       _videoRecordingLocked = false;
     });
 
-    if (mc.cameraController == null || !mc.cameraController!.value.isRecordingVideo) {
+    if (mc.cameraController == null ||
+        !mc.cameraController!.value.isRecordingVideo) {
       return;
     }
 
@@ -638,7 +655,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
 
   @override
   Widget build(BuildContext context) {
-    if (mc.selectedCameraDetails.cameraId >= AppEnvironment.cameras.length || mc.cameraController == null) {
+    if (mc.selectedCameraDetails.cameraId >= AppEnvironment.cameras.length ||
+        mc.cameraController == null) {
       return Container();
     }
     return StreamBuilder(
@@ -662,7 +680,9 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
                 _baseScaleFactor = mc.selectedCameraDetails.scaleFactor;
               });
               // Get the position of the pointer
-              final renderBox = keyTriggerButton.currentContext!.findRenderObject()! as RenderBox;
+              final renderBox =
+                  keyTriggerButton.currentContext!.findRenderObject()!
+                      as RenderBox;
               final localPosition = renderBox.globalToLocal(
                 details.globalPosition,
               );
@@ -698,18 +718,24 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
                       ),
                     ),
                   ),
-                if (!mc.isSharePreviewIsShown && widget.sendToGroup != null && !mc.isVideoRecording)
+                if (!mc.isSharePreviewIsShown &&
+                    widget.sendToGroup != null &&
+                    !mc.isVideoRecording)
                   ShowTitleText(
                     title: widget.sendToGroup!.groupName,
                     desc: context.lang.cameraPreviewSendTo,
                   ),
-                if (!mc.isSharePreviewIsShown && mc.sharedLinkForPreview != null && !mc.isVideoRecording)
+                if (!mc.isSharePreviewIsShown &&
+                    mc.sharedLinkForPreview != null &&
+                    !mc.isVideoRecording)
                   ShowTitleText(
                     title: mc.sharedLinkForPreview?.host ?? '',
                     desc: 'Link',
                     isLink: true,
                   ),
-                if (!mc.isSharePreviewIsShown && !mc.isVideoRecording && !widget.hideControllers)
+                if (!mc.isSharePreviewIsShown &&
+                    !mc.isVideoRecording &&
+                    !widget.hideControllers)
                   CameraTopActions(
                     selectedCameraDetails: mc.selectedCameraDetails,
                     hasAudioPermission: _hasAudioPermission,
@@ -753,7 +779,8 @@ class _CameraPreviewViewState extends State<CameraPreviewView> {
                   videoRecordingStarted: _videoRecordingStarted,
                   maxVideoRecordingTime: maxVideoRecordingTime,
                 ),
-                if (!mc.isSharePreviewIsShown && widget.sendToGroup != null || widget.hideControllers)
+                if (!mc.isSharePreviewIsShown && widget.sendToGroup != null ||
+                    widget.hideControllers)
                   Positioned(
                     left: 5,
                     top: 10,